Move fake rule into main.
