+++ psl-core/src/test/java/org/linqs/psl/reasoner/admm/term/ADMMObjectiveTermTest.java
@@ -19,14 +19,11 @@ package org.linqs.psl.reasoner.admm.term;
 
 import static org.junit.Assert.assertEquals;
 
+import org.linqs.psl.model.rule.FakeRule;
 import org.linqs.psl.reasoner.function.FunctionComparator;
 import org.linqs.psl.reasoner.term.Hyperplane;
 
-import org.junit.Before;
 import org.junit.Test;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class ADMMObjectiveTermTest {
     @Test
